live url

+++ lib/models/order_model.dart
@@ -1,11 +1,11 @@
 class Order {
   int orderId;
-  int orderTotal;
+  double orderTotal;
   String orderStatus;
   String orderCreatedAt;
   Order(json) {
     orderId = json['order_id'];
-    orderTotal = json['total'];
+    orderTotal = json['total'].toDouble();
     orderStatus = json['status'];
     orderCreatedAt = json['created_at'];
   }
